Update when particles are added
Also kinect code is currently commented so one can compile

+++ MusicWall3D/MusicWall3D/ParticleSystem.cs
@@ -24,7 +24,7 @@ namespace ParticleLibrary
 
         }
 
-        public void addParticle(int x, int y)//, ParticleType p, int life)
+        public void addParticle(float x, float y)//, ParticleType p, int life)
         {
             int i = ran.Next(1, 100);
             pSystem.Add(new Glitter(x, y, 200, graphicsDeviceManager, i));
